Separate debug related menu actions into the new debug module.
